Move assertIsBlockhash from transactions to rpc-types
