+++ packages/transactions/src/blockhash.ts
@@ -1,6 +1,4 @@
-import { Encoder } from '@solana/codecs-core';
-import { getBase58Encoder } from '@solana/codecs-strings';
-import type { Blockhash } from '@solana/rpc-types';
+import { assertIsBlockhash, type Blockhash } from '@solana/rpc-types';
 
 import { IDurableNonceTransaction } from './durable-nonce';
 import { ITransactionWithSignatures } from './signatures';
@@ -14,33 +12,6 @@ type BlockhashLifetimeConstraint = Readonly<{
 
 export interface ITransactionWithBlockhashLifetime {
     readonly lifetimeConstraint: BlockhashLifetimeConstraint;
-}
-
-let base58Encoder: Encoder<string> | undefined;
-
-export function assertIsBlockhash(putativeBlockhash: string): asserts putativeBlockhash is Blockhash {
-    if (!base58Encoder) base58Encoder = getBase58Encoder();
-    try {
-        // Fast-path; see if the input string is of an acceptable length.
-        if (
-            // Lowest value (32 bytes of zeroes)
-            putativeBlockhash.length < 32 ||
-            // Highest value (32 bytes of 255)
-            putativeBlockhash.length > 44
-        ) {
-            throw new Error('Expected input string to decode to a byte array of length 32.');
-        }
-        // Slow-path; actually attempt to decode the input string.
-        const bytes = base58Encoder.encode(putativeBlockhash);
-        const numBytes = bytes.byteLength;
-        if (numBytes !== 32) {
-            throw new Error(`Expected input string to decode to a byte array of length 32. Actual length: ${numBytes}`);
-        }
-    } catch (e) {
-        throw new Error(`\`${putativeBlockhash}\` is not a blockhash`, {
-            cause: e,
-        });
-    }
 }
 
 function isTransactionWithBlockhashLifetime(
